requirements.txt
fastapi
uvicorn[standard]
cryptography
python-dotenv
requests
openai
pydantic
replicate
google-generativeai